Upgraded Test Apps to netcoreapp3.0
It had been reported that apps running netcoreapp3.0 had issues loading
librdkafka on Windows Machines. Upgraded the TestApp and TestApi
projects to netcoreapp3.0 to test the issue.

It appears that with netcoreapp3.0 builds, all dependencies including
librdkafka runtimes are copied into the bin directory, much like a
publish, but on windows machines, Confluent.Kafka is expecting the
librdkafka to be in the Global Cache somewhere and gets confused when
they also exist in the bin folder - Not an issue on a linux machine.
Setting CopyLocalLockFileAssemblies to false on a netcoreapp3.0 project
forces only the project files to be output much like netcoreapp2.0/2.2